+++ object_store/src/aws/mod.rs
@@ -29,7 +29,6 @@
 //! [automatic cleanup]: https://aws.amazon.com/blogs/aws/s3-lifecycle-management-update-support-for-multipart-uploads-and-delete-markers/
 
 use async_trait::async_trait;
-use chrono::Utc;
 use futures::stream::BoxStream;
 use futures::{StreamExt, TryStreamExt};
 use reqwest::header::{HeaderName, IF_MATCH, IF_NONE_MATCH};
@@ -315,27 +314,6 @@ impl ObjectStore for AmazonS3 {
             Err(e) => Err(e.into()),
             Ok(_) => Ok(()),
         }
-    }
-
-    async fn delete_prefix(&self, prefix: Option<&Path>, ttl: u64) -> Result<()> {
-        let ttl = chrono::Duration::try_seconds(ttl as i64).unwrap();
-        let locations = self
-            .list(prefix)
-            .try_filter_map(|meta| async move {
-                let cutoff = Utc::now() - ttl;
-                if meta.last_modified < cutoff {
-                    Ok(Some(meta.location))
-                } else {
-                    Ok(None)
-                }
-            })
-            .boxed();
-
-        self.delete_stream(locations)
-            .try_collect::<Vec<Path>>()
-            .await?;
-
-        Ok(())
     }
 }
