+++ object_store/src/gcp/mod.rs
@@ -46,10 +46,8 @@ use crate::{
     UploadPart,
 };
 use async_trait::async_trait;
-use chrono::Utc;
 use client::GoogleCloudStorageClient;
 use futures::stream::BoxStream;
-use futures::{StreamExt, TryStreamExt};
 use hyper::Method;
 use url::Url;
 
@@ -207,27 +205,6 @@ impl ObjectStore for GoogleCloudStorage {
 
     async fn copy_if_not_exists(&self, from: &Path, to: &Path) -> Result<()> {
         self.client.copy_request(from, to, true).await
-    }
-
-    async fn delete_prefix(&self, prefix: Option<&Path>, ttl: u64) -> Result<()> {
-        let ttl = chrono::Duration::try_seconds(ttl as i64).unwrap();
-        let locations = self
-            .list(prefix)
-            .try_filter_map(|meta| async move {
-                let cutoff = Utc::now() - ttl;
-                if meta.last_modified < cutoff {
-                    Ok(Some(meta.location))
-                } else {
-                    Ok(None)
-                }
-            })
-            .boxed();
-
-        self.delete_stream(locations)
-            .try_collect::<Vec<Path>>()
-            .await?;
-
-        Ok(())
     }
 }
