+++ object_store/src/lib.rs
@@ -780,7 +780,26 @@ pub trait ObjectStore: std::fmt::Display + Send + Sync + Debug + 'static {
     }
 
     /// Delete all objects under specified prefix path, otherwise delete all objects
-    async fn delete_prefix(&self, prefix: Option<&Path>, ttl: u64) -> Result<()>;
+    async fn delete_prefix(&self, prefix: Option<&Path>, ttl: u64) -> Result<()> {
+        let ttl = chrono::Duration::try_seconds(ttl as i64).unwrap();
+        let locations = self
+            .list(prefix)
+            .try_filter_map(|meta| async move {
+                let cutoff = Utc::now() - ttl;
+                if meta.last_modified < cutoff {
+                    Ok(Some(meta.location))
+                } else {
+                    Ok(None)
+                }
+            })
+            .boxed();
+
+        self.delete_stream(locations)
+            .try_collect::<Vec<Path>>()
+            .await?;
+
+        Ok(())
+    }
 }
 
 macro_rules! as_ref_impl {
