+++ object_store/src/memory.rs
@@ -396,27 +396,6 @@ impl ObjectStore for InMemory {
         storage.insert(to, entry.data, entry.attributes);
         Ok(())
     }
-
-    async fn delete_prefix(&self, prefix: Option<&Path>, ttl: u64) -> Result<()> {
-        let ttl = chrono::Duration::try_seconds(ttl as i64).unwrap();
-        let locations = self
-            .list(prefix)
-            .try_filter_map(|meta| async move {
-                let cutoff = Utc::now() - ttl;
-                if meta.last_modified < cutoff {
-                    Ok(Some(meta.location))
-                } else {
-                    Ok(None)
-                }
-            })
-            .boxed();
-
-        self.delete_stream(locations)
-            .try_collect::<Vec<Path>>()
-            .await?;
-
-        Ok(())
-    }
 }
 
 #[async_trait]
